compiles with centroid-finding functionality. testing pending.

+++ src/diag_moment.h
@@ -22,6 +22,7 @@ DiagStyle(moment,DiagMoment)
 #include "stdio.h"
 #include "diag.h"
 #include <map>
+#include "grain.h"
 
 namespace SPPARKS_NS {
 
@@ -41,7 +42,7 @@ class DiagMoment : public Diag {
   double min_dist_x(double, double);
   double min_dist_y(double, double);
   double min_dist_z(double, double);
-  void merge_grain(int,int,double,double,double,int,double*);
+  void merge_grain(int,double,double,double,double,int,double*);
   
  protected:
   std::map<int, Grain> grains;
